refactor(layout): reorganize imports and enhance layout structure with Footer component
- Update import paths for AdminSidebar and Navbar components to reflect new directory structure.
- Add Footer component to both Admin and Vault layouts for consistent footer display.
- Improve layout styling for better responsiveness and visual consistency.
