+++ app/(admin)/admin/login/page.tsx
@@ -1,10 +1,9 @@
-import LoginForm from '@/components/auth/login-form';
+import LoginForm from '@/components/layout/auth/login-form';
 
 const AdminLoginPage = () => {
   return (
-    <section>
-      <h1>Login</h1>
-      {/* TODO: Stylen der LoginForm/LoginPage */}
+    <section className="flex flex-col items-center justify-center gap-8 lg:min-h-screen">
+      <h1 className="text-4xl">Login</h1>
       <LoginForm />
     </section>
   );
